accidentally modified facet list, undoing and making better
by using `.delete` to remove a facet from the "home" action,
it was also being deleted from the index page. Therefore, instead
just passing a list of facets that shouldn't be drawn on the home page.

While in there, I took the liberty of adding any facets which were
somehow available on a home page and added them to the search filters